Separate Usability impl
