+++ src/main/kotlin/io/github/freya022/botcommands/internal/commands/text/TextCommandInfoImpl.kt
@@ -1,12 +1,12 @@
 package io.github.freya022.botcommands.internal.commands.text
 
+import io.github.freya022.botcommands.api.commands.Usability.UnusableReason
 import io.github.freya022.botcommands.api.commands.text.TextCommandInfo
 import io.github.freya022.botcommands.api.commands.text.builder.TextCommandBuilder
 import io.github.freya022.botcommands.api.core.utils.toImmutableList
 import io.github.freya022.botcommands.api.core.utils.unmodifiableView
 import io.github.freya022.botcommands.internal.commands.AbstractCommandInfoImpl
-import io.github.freya022.botcommands.internal.commands.Usability
-import io.github.freya022.botcommands.internal.commands.Usability.UnusableReason
+import io.github.freya022.botcommands.internal.commands.UsabilityImpl
 import io.github.freya022.botcommands.internal.utils.putIfAbsentOrThrow
 import io.github.freya022.botcommands.internal.utils.throwInternal
 import net.dv8tion.jda.api.EmbedBuilder
@@ -50,7 +50,7 @@ internal sealed class TextCommandInfoImpl(
         }
     }
 
-    final override fun getUsability(member: Member, channel: GuildMessageChannel) = Usability.build {
+    final override fun getUsability(member: Member, channel: GuildMessageChannel) = UsabilityImpl.build {
         val isNotOwner = !context.isOwner(member.idLong)
         if (isNotOwner && hidden) add(UnusableReason.HIDDEN)
         if (isNotOwner && isOwnerRequired) add(UnusableReason.OWNER_ONLY)
